Add item indicators to group items in the rundown
